requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
pyparsing==2.3.0
pyspeedtest==1.2.7
python-dateutil==2.7.5
python-nmap==0.6.1
six==1.11.0
